Test for possible Case switch
Modified series of if/elif statements into a case switch. Possible refactoring of code to reduce overhead, improve readability, and provide ease of maintenance. Also changed the base case to Bluehost. Since the variable is used with a database, default case of Bluehost ensures that a brand is provided at all times.