Modify Branding Case Switch
The original code required the user to explicitly type in the brand, with specific capitalization. For lazy users, like myself, who types in all lowercase, this resulted in the "Generic branding set". Changed the case switch to convert the input to all lowercase, and then compare. When passing the brand, the proper capitalization is then preserved.